Changes to SummaryDetail for the slide in/out effect
Since SummaryDetail is now always rendered in App.jsx, recipeDetail is only destructured if it exists. Also, the container with the recipe information is only rendered if recipeDetail exists. The styles of this container are set based on isDetailShown, to create the transform and transition properties necessary for the slide in/out effect